Refactor(activity): 활동 게시판 관련 컴포넌트 구조 수정
