+++ src/app/(main)/activity/[semesterName]/[activityId]/boards/[boardId]/_components/board/index.ts
@@ -1,0 +1,1 @@
+export { BoardHero } from './BoardHero'
