+++ src/app/(main)/activity/[semesterName]/[activityId]/boards/[boardId]/_components/index.ts
@@ -1,0 +1,2 @@
+export * from './board'
+export * from './post'
