+++ src/app/(main)/activity/[semesterName]/[activityId]/boards/[boardId]/_components/post/CreateButton.tsx
@@ -1,7 +1,5 @@
 'use client'
 
-import { useEffect, useState } from 'react'
-
 import { usePathname, useRouter } from 'next/navigation'
 
 import { Button } from '@/components/ui/button'
@@ -13,17 +11,12 @@ export const CreatePostButton = () => {
   const router = useRouter()
 
   const { role } = useMyInfoStore((state) => state.getMyInfo())
-  const [isAble, setisAble] = useState(false)
-
-  useEffect(() => {
-    setisAble(role?.includes(role as Role) ?? false)
-  }, [role])
 
   return (
     <div className="mb-20 flex justify-end">
       <Button
         onClick={() => router.push(`${pathName}/create-post`)}
-        disabled={!isAble}
+        disabled={!role?.includes(role as Role)}
         className="max-w-fit"
       >
         게시글 생성하기
