+++ src/app/(main)/activity/[semesterName]/[activityId]/boards/[boardId]/_components/post/PostList.tsx
