+++ src/app/(main)/activity/[semesterName]/[activityId]/boards/[boardId]/_components/post/index.ts
@@ -1,0 +1,2 @@
+export { ActivityPostListSection } from './PostList'
+export { CreatePostButton } from './CreateButton'
