--- src/app/(main)/activity/[semesterName]/[activityId]/boards/[boardId]/create-post/_components/CreateActivityPostHero/index.tsx
@@ -1,42 +1,0 @@
-'use client'
-
-import { useQuery } from '@tanstack/react-query'
-import { usePathname } from 'next/navigation'
-
-import { NavLink } from '@/components/PageBreadcrumb'
-import { ActivityBreadcrumb } from '@/components/feature'
-import { Separator } from '@/components/ui/separator'
-import { DATA_ERROR_MESSAGES } from '@/constant/errorMessage'
-import { boardDetailQuery } from '@/service/data/boards'
-
-type CreateActivityPostHeroProps = {
-  activityId: number
-  boardId: number
-}
-
-export const CreateActivityPostHero = ({
-  activityId,
-  boardId,
-}: CreateActivityPostHeroProps) => {
-  const pathName = usePathname()
-
-  const boardPath = pathName.split('/').slice(0, -1).join('/')
-  const { data: boardDetail } = useQuery(boardDetailQuery(activityId, boardId))
-
-  if (!boardDetail) throw new Error(DATA_ERROR_MESSAGES.BOARD_DETAIL_NOT_FOUND)
-
-  const navLinks: NavLink[] = [
-    {
-      link: `${boardPath}`,
-      name: `${boardDetail.boardName} 게시판`,
-    },
-  ]
-
-  return (
-    <div>
-      <Separator variant="dark" />
-      <ActivityBreadcrumb navLinks={navLinks} pageName="게시글 생성하기" />
-      <Separator variant="dark" />
-    </div>
-  )
-}
