+++ src/components/feature/board/BoardDetail.tsx
