+++ src/components/feature/skeleton/BoardHero.tsx
