+++ src/components/feature/skeleton/index.ts
@@ -1,3 +1,4 @@
 export { ActivitySkeleton } from './Activity'
 export { AdminSemesterSkeleton, ActivitySemesterSkeleton } from './Semester'
 export { BoardSkeleton } from './Board'
+export { BoardHeroSkeleton } from './BoardHero'
